setting up deleting method form instances to instance_id file and collections to collection_id file
