+++ actinia_stac_plugin/api/stac_collection_id.py
@@ -23,7 +23,10 @@ __maintainer__ = "__mundialis__"
 from actinia_core.rest.resource_base import ResourceBase
 from flask import make_response
 
-from actinia_stac_plugin.core.stac_collection_id import callStacCollection
+from actinia_stac_plugin.core.stac_collection_id import (
+    callStacCollection,
+    deleteStacCollection,
+)
 
 
 class StacCollections(ResourceBase):
@@ -33,8 +36,20 @@ class StacCollections(ResourceBase):
         ResourceBase.__init__(self)
 
     # @swagger.doc(modules.listModules_get_docs)
-    def get(self, stac_collection_id):
+    def get(self, stac_collection_id: str):
         """Get a list of specified Collection."""
 
         collection_list = callStacCollection(stac_collection_id)
         return make_response(collection_list, 200)
+
+    def delete(self, stac_collection_id: str):
+        """
+            This function delete the STAC Catalog stored before on ID basis.
+            Arg:
+                - ID - ID/Name given to the STAC Catalog you want to delete
+        """
+        stac_instance_id = stac_collection_id.split(".")[1]
+
+        deleted_stac = deleteStacCollection(stac_instance_id, stac_collection_id)
+
+        return make_response(deleted_stac, 200)
