--- actinia_stac_plugin/model/responseModels.py
@@ -1,46 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""
-Copyright (c) 2021 mundialis GmbH & Co. KG
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-
-Common api methods
-"""
-
-__author__ = "Carmen Tawalika"
-__copyright__ = "2018-2021 mundialis GmbH & Co. KG"
-__license__ = "GPLv3"
-
-
-from flask_restful_swagger_2 import Schema
-
-
-class SimpleStatusCodeResponseModel(Schema):
-    """Simple response schema to inform about status."""
-
-    type = "object"
-    properties = {
-        "status": {"type": "number", "description": "The status code of the request."},
-        "message": {
-            "type": "string",
-            "description": "A short message to describes the status",
-        },
-    }
-    required = ["status", "message"]
-
-
-simpleResponseExample = SimpleStatusCodeResponseModel(status=200, message="success")
-SimpleStatusCodeResponseModel.example = simpleResponseExample
